Bouncing and Levels
I've decided I'm going to design levels rather than generate them.

So I've started some of that here! We have a level, it moves up, and the
plane bumps into the things and stops. And when it bumps into something
going downward, the level bumps upwards so you can't get dragged off the
bottom.

The plane knows it hit something, the wall knows it hit the plane, I
think we're all good here for now.

There's a bit of a physics bug when the timescale is set pretty high
where the plane doesn't notice being pushed when it's not trying to
move, but that's probably ok to ignore for now.
It's unplayably fast.