Basic Health System and Throttle
I can now change the speed of the plane, because it was really boring
before.
This lead to me running into the bug I mentioned in the last commit
message where I'd get pushed if I wasn't moving...

That was inconvenient, and I had to put in a hack to fix it.

Beyond that I also made a basic health system. I lose health when I bump
into stuff, and I can heal with dice.

But, too many health and it blows and I'm back to 1! Oh no!