Basic UI
Holy shit did that take longer than expected...
I've got a main screen, a resolution chosen, and some fancy gauges and
stuff.

I rebuilt the dice mechanics from the ground up, and totally redid the
shield UI, and actually made a throttle UI.

So after 5 hours, I'm in exactly the same place I was before, but this
time prettier.

I hoped to go to bed tonight with a done game with only one boring game,
but instead I just have a few pictures and just as little of a game as I
did in mid-day.

I hope tomorrow goes better!