Affine Transform
Ok, so I think this was the reason things weren't transforming right.

Apparently there's a different transform when scaling is involved!

Fixes the top left point in browser, but still not bottom right...
I even put a picture there and it looked right.
I wonder if it moved after on_ready?