Level Loader, Levels, and Release
This was a cram and so I didn't commit things.
So here it all is!
I built a level-loader and checkpoint system at the last minute so you
can fly through stuff and when you die you don't go all the way to the
back.

I also coloured the mountains the same as the ground by just editing the
png directly.

There's still some things that grump me, but this is where I got to in
the time alotted.